refactor(index): move IDE index construction to indexing layer
