+++ crates/ide/src/lib.rs
@@ -28,6 +28,7 @@ pub mod formatting;
 pub mod goto_declaration;
 pub mod goto_definition;
 pub mod hover;
+mod indexing;
 pub mod inlay_hint;
 #[cfg(test)]
 mod macro_hover_tests;
